chore(cli): split cli_tests.rs to satisfy the 700-line pre-push gate
src/cli_tests.rs sat at 705 lines, past the repo's linecheck gate in
.githooks/pre-push, so a clean checkout of main currently fails pre-push
for anyone with the git hooks installed. Move the 7 BIND_ADDR_ENV-override
tests into a new cli_bind_override_tests.rs sibling, with its own EnvGuard
copy per the existing one-helper-copy-per-test-file convention. No test
bodies changed.
